feat: allow image draws with stacked mask texture
DrawImage and DrawImageQuad now supports a final maskIdx that indicates
which texture in the stack to multiply the final output with; adding the
ability to mask out undesired parts of the image.

As it needs to be in the same stack, the mask image needs to have the
same size and image format as the colour image.

+++ engine/render/r_texture.cpp
@@ -446,7 +446,7 @@ std::unique_ptr<image_c> r_tex_c::BuildMipSet(std::unique_ptr<image_c> img)
 				const size_t levels = newTex.levels();
 				for (size_t level = 1; level < levels; ++level) {
 					const auto srcExtent = newTex.extent(level - 1);
-					const auto comp = gli::component_count(format);
+					const auto comp = (int)gli::component_count(format);
 					const auto dstExtent = newTex.extent(level);
 					const bool hasAlpha = comp == 4;
 					stbir_resize_uint8_srgb_edgemode(
